Add `DropNotes` transform to filter out random notes.
This requires a minor change to the protocol for transforms:
`transform(message)` may now return `None`, meaning "delete that message".

It was surprisingly tricky to get this right because of the way the
`.time` attribute of a MIDI message is only a time *delta* (how long to
sleep before playing the note, basically) rather than a timestamp.